added splash screen
