+++ src/main/res/values/themes.xml
@@ -160,4 +160,8 @@
         <item name="android:windowBackground">@android:color/black</item>
     </style>
 
+    <style name="SplashTheme" parent="Theme.AppCompat.NoActionBar">
+        <item name="colorPrimaryDark">@color/grey50</item>
+        <item name="android:windowBackground">@drawable/background</item>
+    </style>
 </resources>
